// ui/src/types.ts
export interface Certificate {
  instituteLogo?: string;
  instituteName: string;
  certificateType: string;
  studentName: string;
  courseName: string;
  specialization: string;
  grade: string;
  startDate: string;
  endDate: string;
  achievements: string;
  authoritySignature?: string;
  certifyingAuthority: string;
  authorityDesignation: string;
  certificateId: string;
}